login + register

+++ backend/controllers/todoos.js
@@ -1,26 +1,44 @@
 const todooRouter = require("express").Router();
-const { where } = require("sequelize");
-const Todoo = require("../models/todoos");
+const { Todoo } = require("../models/index");
+const middleware = require("../utils/middleware");
+
+todooRouter.use(middleware.authenticateToken);
 
 todooRouter.post("/", async (request, response) => {
+  const title = request.body.title;
+  const description = request.body.description;
+  const isCompleted = request.body.isCompleted;
+  const category = request.body.category;
+  const user_id = request.UserId;
   try {
-    const todoo = await Todoo.create(request.body);
+    const todoo = await Todoo.create({
+      title,
+      description,
+      isCompleted,
+      category,
+      user_id,
+    });
     return response.json(todoo);
   } catch (error) {
-    response.status(400).json({ error });
+    return response.status(400).json({ error });
   }
 });
 
 todooRouter.get("/", async (request, response) => {
-  const todoos = await Todoo.findAll();
+  const UserId = request.UserId;
+  const todoos = await Todoo.findAll({
+    where: { user_id: UserId },
+    attributes: { exclude: ["userId"] },
+  });
   response.json(todoos);
 });
 
 todooRouter.put("/:id", async (request, response) => {
+  const user_id = request.UserId;
   const id = Number(request.params.id);
   try {
     const todoo = await Todoo.update(request.body, {
-      where: { id: id },
+      where: { id: id, user_id: user_id },
       returning: true,
     });
     response.json(todoo[1][0]);
@@ -30,9 +48,10 @@ todooRouter.put("/:id", async (request, response) => {
 });
 
 todooRouter.patch("/:id", async (request, response) => {
+  const user_id = request.UserId;
   const id = Number(request.params.id);
   const todoo = await Todoo.update(request.body, {
-    where: { id: id },
+    where: { id: id, user_id: user_id },
     returning: true,
   });
   response.json(todoo[1][0]);
@@ -40,8 +59,9 @@ todooRouter.patch("/:id", async (request, response) => {
 
 todooRouter.delete("/:id", (request, response) => {
   const id = Number(request.params.id);
+  const user_id = request.UserId;
   try {
-    Todoo.destroy({ where: { id: id } });
+    Todoo.destroy({ where: { id: id, user_id: user_id } });
     response.status(204).end();
   } catch {
     response.status(400).json({ error });
